refactor: casedb.services.case moved crud methods to separate module
