+++ gen_epix/casedb/services/case/read_association_with_valid_ids.py
@@ -2,7 +2,7 @@ from typing import Type
 from uuid import UUID
 
 from gen_epix.casedb.domain import command, model
-from gen_epix.casedb.domain.service import BaseCaseService
+from gen_epix.casedb.services.case.base import BaseCaseService
 from gen_epix.fastapp import BaseUnitOfWork, CrudOperation
 from gen_epix.filter import CompositeFilter, Filter, LogicalOperator, UuidSetFilter
