+++ gen_epix/casedb/services/case/retrieve_phylogenetic_tree.py
@@ -2,7 +2,7 @@ from uuid import UUID
 
 from gen_epix.casedb.domain import command, enum, exc, model
 from gen_epix.casedb.domain.policy.abac import BaseCaseAbacPolicy
-from gen_epix.casedb.domain.service.case import BaseCaseService
+from gen_epix.casedb.services.case.base import BaseCaseService
 from gen_epix.fastapp.enum import CrudOperation
