+++ test/casedb/integration/case_access/test_case_access.py
@@ -88,6 +88,7 @@ class TestCaseAccess(CaseAccessSetup):
         df = env.props["case_crud_command"]
         if df is None:
             raise ValueError("Case CRUD commands DataFrame is not set.")
+        df = df.loc[df["dm.is_active"] == True, :]
         command_idx_to_test = None
         # command_idx_to_test = {6}  # For debugging, set set of indices, otherwise None
         n_case_type_cols = 3
